+++ packages/frontier-plugins/plugin-vue/src/commands/vue/add/service.ts
@@ -1,10 +1,10 @@
 import { Command, flags } from '@oclif/command';
 import chalk from 'chalk';
-import { toKebabCase } from '@rdfrontier/stdlib';
+import { toKebabCase } from '@jantaeleckie/stdlib';
 import { checkProjectValidity, parseServiceName } from '../../../utils/utilities';
 import { CLI_COMMANDS, CLI_STATE, DOCUMENTATION_LINKS } from '../../../utils/constants';
-import { catchError } from '@rdfrontier/plugin-shared';;
-import { invalidProject } from '@rdfrontier/plugin-shared';
+import { catchError } from '@jantaeleckie/plugin-shared';;
+import { invalidProject } from '@jantaeleckie/plugin-shared';
 import { addElementFunction } from '../../../functions/addElement';
 
 const TEMPLATE_FOLDERS = ['service'];