+++ packages/frontier-plugins/plugin-qa/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @rdfrontier/plugin-qa
+# @jantaeleckie/plugin-qa
 
 ## 0.0.3
 
 ### Patch Changes
 
 - Updated dependencies [286d4e6]
-  - @rdfrontier/stdlib@0.0.2
+  - @jantaeleckie/stdlib@0.0.2